Add KeyFrame Skeleton Animation and Animation Speed Feature, Add Example Ranger Game

+++ zener-core/src/com/n8lm/zener/animation/Animation.java
@@ -30,7 +30,7 @@ public class Animation<T extends KeyFrame> {
 
 	protected String name;
 	protected List<T> frames;
-	protected int totalTime;
+	protected float totalTime;
 	
 	public Animation() {
 		this.name = "undefine";
@@ -60,7 +60,7 @@ public class Animation<T extends KeyFrame> {
 		return frames.get(index);
 	}
 	
-	public int getTotalTime() {
+	public float getTotalTime() {
 		return totalTime;
 	}
